refactor: split lib.rs into focused modules
- Separate 800+ line lib.rs into 6 focused modules:
  - error.rs: TylError enum + constructors
  - category.rs: ErrorCategory trait + built-in categories
  - context.rs: ErrorContext struct + builder
  - retry.rs: RetryPolicy + RetryableError trait
  - settings.rs: ErrorSettings + environment variables
  - lib.rs: Re-exports + documentation + tests

- Maintain full backward compatibility
- All 19 tests passing (16 unit + 3 integration + 3 doc)
- Improved maintainability and code organization
