+++ src/lib.rs
@@ -94,9 +94,24 @@
 //! assert!(error.category().is_retriable());
 //! ```
 
+// Module declarations
+mod category;
+mod context;
+mod error;
+mod retry;
+mod settings;
+
+// Re-export main types and traits
+pub use error::{TylError, TylResult};
+pub use category::{ErrorCategory, ErrorClassifier, BuiltinCategory};
+pub use context::ErrorContext;
+pub use retry::{RetryableError, RetryPolicy, RetryResult};
+pub use settings::{ErrorSettings, LogLevel};
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::time::Duration;
 
     #[test]
     fn test_tyl_error_creation_should_create_different_error_types() {
@@ -337,464 +352,5 @@ mod tests {
         // Then: should use custom logic
         assert!(!category.is_retriable());
         assert_eq!(category.category_name(), "BusinessLogic");
-    }
-}
-
-// Implementation will go here - starting with failing tests (TDD red phase)
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::time::Duration;
-use thiserror::Error;
-use uuid::Uuid;
-
-/// Log level for error output
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum LogLevel {
-    Error = 0,
-    Warn = 1,
-    Info = 2,
-    Debug = 3,
-}
-
-impl LogLevel {
-    fn from_env() -> Option<Self> {
-        std::env::var("TYL_ERROR_LOG_LEVEL").ok().and_then(|level| {
-            match level.to_uppercase().as_str() {
-                "ERROR" => Some(LogLevel::Error),
-                "WARN" | "WARNING" => Some(LogLevel::Warn),
-                "INFO" => Some(LogLevel::Info),
-                "DEBUG" => Some(LogLevel::Debug),
-                _ => None,
-            }
-        })
-    }
-}
-
-/// Global error configuration from environment variables
-/// Zero-config with sensible defaults to avoid circular dependencies
-struct ErrorSettings {
-    backtrace_enabled: bool,
-    max_retries: usize,
-    log_errors: bool,
-    log_level: LogLevel,
-}
-
-impl ErrorSettings {
-    /// Get global error settings from environment variables
-    fn global() -> &'static Self {
-        use std::sync::OnceLock;
-        static SETTINGS: OnceLock<ErrorSettings> = OnceLock::new();
-
-        SETTINGS.get_or_init(|| {
-            let backtrace_enabled = std::env::var("TYL_ERROR_BACKTRACE")
-                .map(|v| v.to_lowercase() == "true")
-                .unwrap_or_else(|_| {
-                    // Fall back to RUST_BACKTRACE if TYL_ERROR_BACKTRACE not set
-                    std::env::var("RUST_BACKTRACE").is_ok()
-                });
-
-            let max_retries = std::env::var("TYL_ERROR_MAX_RETRIES")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(3);
-
-            let log_errors = std::env::var("TYL_ERROR_LOG_ERRORS")
-                .map(|v| v.to_lowercase() != "false")
-                .unwrap_or(true);
-
-            let log_level = LogLevel::from_env().unwrap_or(LogLevel::Info);
-
-            ErrorSettings {
-                backtrace_enabled,
-                max_retries,
-                log_errors,
-                log_level,
-            }
-        })
-    }
-}
-
-pub type TylResult<T> = Result<T, TylError>;
-
-/// Trait for defining custom error classification behavior.
-///
-/// This trait allows users to define domain-specific error categories
-/// without modifying the core tyl-errors module.
-pub trait ErrorClassifier: std::fmt::Debug + Send + Sync {
-    /// Determine if this error category should trigger retries.
-    fn is_retriable(&self) -> bool;
-
-    /// Calculate the suggested retry delay for this error category.
-    fn retry_delay(&self, attempt: usize) -> Duration;
-
-    /// Get a human-readable name for this error category.
-    fn category_name(&self) -> &'static str;
-
-    /// Clone this error classifier (needed for ErrorCategory cloning).
-    fn clone_box(&self) -> Box<dyn ErrorClassifier>;
-}
-
-impl Clone for Box<dyn ErrorClassifier> {
-    fn clone(&self) -> Self {
-        self.clone_box()
-    }
-}
-
-// Default classifier for deserialization fallback
-fn default_classifier() -> Box<dyn ErrorClassifier> {
-    Box::new(BuiltinCategory::Unknown)
-}
-
-#[derive(Error, Debug, Clone, Serialize, Deserialize)]
-pub enum TylError {
-    #[error("Database error: {message}")]
-    Database { message: String },
-
-    #[error("Network error: {message}")]
-    Network { message: String },
-
-    #[error("Validation error: {field}: {message}")]
-    Validation { field: String, message: String },
-
-    #[error("Not found: {resource} with id {id}")]
-    NotFound { resource: String, id: String },
-
-    #[error("Conflict: {message}")]
-    Conflict { message: String },
-
-    #[error("Internal error: {message}")]
-    Internal { message: String },
-
-    #[error("Configuration error: {message}")]
-    Configuration { message: String },
-
-    #[error("Feature not implemented: {feature}")]
-    NotImplemented { feature: String },
-
-    #[error("Custom error: {message}")]
-    Custom {
-        message: String,
-        #[serde(skip)]
-        #[serde(default = "default_classifier")]
-        classifier: Box<dyn ErrorClassifier>,
-    },
-}
-
-/// Built-in error categories provided by tyl-errors.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub enum BuiltinCategory {
-    Transient,
-    Permanent,
-    ResourceExhaustion,
-    Network,
-    Authentication,
-    Validation,
-    Internal,
-    ServiceUnavailable,
-    Unknown,
-}
-
-/// Extensible error category system.
-///
-/// Supports both built-in categories and custom user-defined categories.
-#[derive(Debug, Clone)]
-pub enum ErrorCategory {
-    /// Built-in error categories with predefined behavior.
-    Builtin(BuiltinCategory),
-    /// Custom error categories defined by users.
-    Custom(Box<dyn ErrorClassifier>),
-}
-
-impl ErrorClassifier for BuiltinCategory {
-    fn is_retriable(&self) -> bool {
-        matches!(
-            self,
-            BuiltinCategory::Transient
-                | BuiltinCategory::Network
-                | BuiltinCategory::ServiceUnavailable
-                | BuiltinCategory::ResourceExhaustion
-        )
-    }
-
-    fn retry_delay(&self, attempt: usize) -> Duration {
-        let base_delay = match self {
-            BuiltinCategory::Transient => Duration::from_millis(100),
-            BuiltinCategory::Network => Duration::from_millis(500),
-            BuiltinCategory::ServiceUnavailable => Duration::from_secs(1),
-            BuiltinCategory::ResourceExhaustion => Duration::from_secs(5),
-            _ => Duration::from_millis(100),
-        };
-
-        let multiplier = 2_u32.pow(attempt.min(10) as u32);
-        base_delay * multiplier.min(60)
-    }
-
-    fn category_name(&self) -> &'static str {
-        match self {
-            BuiltinCategory::Transient => "Transient",
-            BuiltinCategory::Permanent => "Permanent",
-            BuiltinCategory::ResourceExhaustion => "ResourceExhaustion",
-            BuiltinCategory::Network => "Network",
-            BuiltinCategory::Authentication => "Authentication",
-            BuiltinCategory::Validation => "Validation",
-            BuiltinCategory::Internal => "Internal",
-            BuiltinCategory::ServiceUnavailable => "ServiceUnavailable",
-            BuiltinCategory::Unknown => "Unknown",
-        }
-    }
-
-    fn clone_box(&self) -> Box<dyn ErrorClassifier> {
-        Box::new(self.clone())
-    }
-}
-
-// Convenience constructors for built-in categories
-impl ErrorCategory {
-    pub fn transient() -> Self {
-        Self::Builtin(BuiltinCategory::Transient)
-    }
-    pub fn permanent() -> Self {
-        Self::Builtin(BuiltinCategory::Permanent)
-    }
-    pub fn resource_exhaustion() -> Self {
-        Self::Builtin(BuiltinCategory::ResourceExhaustion)
-    }
-    pub fn network() -> Self {
-        Self::Builtin(BuiltinCategory::Network)
-    }
-    pub fn authentication() -> Self {
-        Self::Builtin(BuiltinCategory::Authentication)
-    }
-    pub fn validation() -> Self {
-        Self::Builtin(BuiltinCategory::Validation)
-    }
-    pub fn internal() -> Self {
-        Self::Builtin(BuiltinCategory::Internal)
-    }
-    pub fn service_unavailable() -> Self {
-        Self::Builtin(BuiltinCategory::ServiceUnavailable)
-    }
-    pub fn unknown() -> Self {
-        Self::Builtin(BuiltinCategory::Unknown)
-    }
-
-    // Delegate methods to the classifier
-    pub fn is_retriable(&self) -> bool {
-        match self {
-            ErrorCategory::Builtin(builtin) => builtin.is_retriable(),
-            ErrorCategory::Custom(custom) => custom.is_retriable(),
-        }
-    }
-
-    pub fn retry_delay(&self, attempt: usize) -> Duration {
-        match self {
-            ErrorCategory::Builtin(builtin) => builtin.retry_delay(attempt),
-            ErrorCategory::Custom(custom) => custom.retry_delay(attempt),
-        }
-    }
-
-    pub fn category_name(&self) -> &str {
-        match self {
-            ErrorCategory::Builtin(builtin) => builtin.category_name(),
-            ErrorCategory::Custom(custom) => custom.category_name(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ErrorContext {
-    pub error_id: Uuid,
-    pub operation: String,
-    #[serde(skip)]
-    #[serde(default = "ErrorCategory::unknown")]
-    pub category: ErrorCategory,
-    pub message: String,
-    pub occurred_at: DateTime<Utc>,
-    pub attempt_count: usize,
-    pub metadata: HashMap<String, serde_json::Value>,
-}
-
-impl TylError {
-    pub fn database<S: Into<String>>(message: S) -> Self {
-        Self::Database {
-            message: message.into(),
-        }
-    }
-
-    pub fn network<S: Into<String>>(message: S) -> Self {
-        Self::Network {
-            message: message.into(),
-        }
-    }
-
-    pub fn validation<F: Into<String>, M: Into<String>>(field: F, message: M) -> Self {
-        Self::Validation {
-            field: field.into(),
-            message: message.into(),
-        }
-    }
-
-    pub fn not_found<R: Into<String>, I: Into<String>>(resource: R, id: I) -> Self {
-        Self::NotFound {
-            resource: resource.into(),
-            id: id.into(),
-        }
-    }
-
-    pub fn conflict<S: Into<String>>(message: S) -> Self {
-        Self::Conflict {
-            message: message.into(),
-        }
-    }
-
-    pub fn internal<S: Into<String>>(message: S) -> Self {
-        Self::Internal {
-            message: message.into(),
-        }
-    }
-
-    pub fn configuration<S: Into<String>>(message: S) -> Self {
-        Self::Configuration {
-            message: message.into(),
-        }
-    }
-
-    pub fn not_implemented<S: Into<String>>(feature: S) -> Self {
-        Self::NotImplemented {
-            feature: feature.into(),
-        }
-    }
-
-    pub fn business_logic<S: Into<String>>(
-        message: S,
-        classifier: Box<dyn ErrorClassifier>,
-    ) -> Self {
-        Self::Custom {
-            message: message.into(),
-            classifier,
-        }
-    }
-
-    pub fn category(&self) -> ErrorCategory {
-        match self {
-            TylError::Database { .. } => ErrorCategory::transient(),
-            TylError::Network { .. } => ErrorCategory::network(),
-            TylError::Validation { .. } => ErrorCategory::validation(),
-            TylError::NotFound { .. } => ErrorCategory::permanent(),
-            TylError::Conflict { .. } => ErrorCategory::permanent(),
-            TylError::Internal { .. } => ErrorCategory::internal(),
-            TylError::Configuration { .. } => ErrorCategory::permanent(),
-            TylError::NotImplemented { .. } => ErrorCategory::permanent(),
-            TylError::Custom { classifier, .. } => ErrorCategory::Custom(classifier.clone()),
-        }
-    }
-
-    pub fn to_context(&self, operation: String) -> ErrorContext {
-        ErrorContext::new(operation, self.category(), self.to_string())
-    }
-
-    // Convenience methods
-    pub fn parsing<S: Into<String>>(message: S) -> Self {
-        Self::Validation {
-            field: "parsing".to_string(),
-            message: message.into(),
-        }
-    }
-
-    pub fn serialization<S: Into<String>>(message: S) -> Self {
-        let msg = message.into();
-        Self::Internal {
-            message: format!("Serialization error: {msg}"),
-        }
-    }
-
-    pub fn connection<S: Into<String>>(message: S) -> Self {
-        let msg = message.into();
-        Self::Network {
-            message: format!("Connection error: {msg}"),
-        }
-    }
-
-    pub fn initialization<S: Into<String>>(message: S) -> Self {
-        let msg = message.into();
-        Self::Internal {
-            message: format!("Initialization error: {msg}"),
-        }
-    }
-
-    /// Check if backtraces are enabled via environment variables
-    ///
-    /// Checks TYL_ERROR_BACKTRACE first, falls back to RUST_BACKTRACE
-    pub fn backtrace_enabled() -> bool {
-        ErrorSettings::global().backtrace_enabled
-    }
-
-    /// Get maximum retry attempts from TYL_ERROR_MAX_RETRIES (default: 3)
-    pub fn max_retries() -> usize {
-        ErrorSettings::global().max_retries
-    }
-
-    /// Check if error logging is enabled via TYL_ERROR_LOG_ERRORS (default: true)
-    pub fn log_errors_enabled() -> bool {
-        ErrorSettings::global().log_errors
-    }
-
-    /// Get current log level from TYL_ERROR_LOG_LEVEL (default: INFO)
-    pub fn log_level() -> LogLevel {
-        ErrorSettings::global().log_level
-    }
-
-    /// Check if this error should be retried based on attempt count and max retries
-    pub fn should_retry(&self, attempt: usize) -> bool {
-        self.category().is_retriable() && attempt < Self::max_retries()
-    }
-
-    /// Log error if logging is enabled and meets log level criteria
-    pub fn log_if_enabled(&self, level: LogLevel) {
-        if Self::log_errors_enabled() && level <= Self::log_level() {
-            eprintln!(
-                "[{}] {}",
-                match level {
-                    LogLevel::Error => "ERROR",
-                    LogLevel::Warn => "WARN",
-                    LogLevel::Info => "INFO",
-                    LogLevel::Debug => "DEBUG",
-                },
-                self
-            );
-        }
-    }
-}
-
-// From implementations for common error types
-impl From<serde_json::Error> for TylError {
-    fn from(err: serde_json::Error) -> Self {
-        Self::Internal {
-            message: format!("JSON serialization error: {err}"),
-        }
-    }
-}
-
-impl ErrorContext {
-    pub fn new(operation: String, category: ErrorCategory, message: String) -> Self {
-        Self {
-            error_id: Uuid::new_v4(),
-            operation,
-            category,
-            message,
-            occurred_at: Utc::now(),
-            attempt_count: 1,
-            metadata: HashMap::new(),
-        }
-    }
-
-    pub fn with_metadata(mut self, key: String, value: serde_json::Value) -> Self {
-        self.metadata.insert(key, value);
-        self
-    }
-
-    pub fn increment_attempt(&mut self) {
-        self.attempt_count += 1;
     }
 }
